220: deprecating obsolete Quantity implementations
