+++ src/main/java/tech/units/indriya/quantity/deprecated/LongQuantity.java
@@ -27,10 +27,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tech.units.indriya.quantity;
+package tech.units.indriya.quantity.deprecated;
 
 import java.math.BigDecimal;
-
 import javax.measure.Quantity;
 import javax.measure.Unit;
 
@@ -38,53 +37,54 @@ import tech.units.indriya.AbstractQuantity;
 import tech.units.indriya.ComparableQuantity;
 
 /**
- * An amount of quantity, consisting of a short and a Unit. ByteQuantity objects are immutable.
+ * An amount of quantity, consisting of a long and a Unit. LongQuantity objects are immutable.
  * 
  * @see AbstractQuantity
  * @see Quantity
- * @author <a href="mailto:units@catmedia.us">Werner Keil</a>
+ * @author <a href="mailto:werner@uom.technology">Werner Keil</a>
  * @param <Q>
  *            The type of the quantity.
- * @version 0.3, $Date: 2018-10-31 $
+ * @version 0.5, $Date: 2018-10-31 $
  * @since 1.0
  */
-final class ByteQuantity<Q extends Quantity<Q>> extends JavaNumericQuantity<Q> {
+@Deprecated
+final class LongQuantity<Q extends Quantity<Q>> extends JavaNumericQuantity<Q> {
 
-    private static final long serialVersionUID = 6325849816534488248L;
+    private static final long serialVersionUID = 3092808554937634365L;
 
-    private static final BigDecimal BYTE_MIN_VALUE = new BigDecimal(Byte.MIN_VALUE);
-    private static final BigDecimal BYTE_MAX_VALUE = new BigDecimal(Byte.MAX_VALUE);
+    private static final BigDecimal LONG_MAX_VALUE = new BigDecimal(Long.MAX_VALUE);
+    private static final BigDecimal LONG_MIN_VALUE = new BigDecimal(Long.MIN_VALUE);
 
-    private final byte value;
+    private final long value;
 
-    ByteQuantity(byte value, Unit<Q> unit, Scale sc) {
+    LongQuantity(long value, Unit<Q> unit,  Scale sc) {
         super(unit, sc);
         this.value = value;
     }
 
-    ByteQuantity(byte value, Unit<Q> unit) {
+    LongQuantity(long value, Unit<Q> unit) {
         super(unit);
         this.value = value;
     }
 
     @Override
-    public Byte getValue() {
+    public Long getValue() {
         return value;
+    }
+
+    @Override
+    boolean isOverflowing(BigDecimal aValue) {
+        return aValue.compareTo(LONG_MIN_VALUE) < 0 || aValue.compareTo(LONG_MAX_VALUE) > 0;
+    }
+
+    @SuppressWarnings({ "rawtypes", "unchecked" })
+    public ComparableQuantity<Q> inverse() {
+        return new LongQuantity(1 / value, getUnit().inverse());
     }
 
     @Override
     public boolean isBig() {
         return false;
-    }
-
-    @Override
-    boolean isOverflowing(BigDecimal aValue) {
-        return aValue.compareTo(BYTE_MIN_VALUE) < 0 || aValue.compareTo(BYTE_MAX_VALUE) > 0;
-    }
-
-    @Override
-    public ComparableQuantity<?> inverse() {
-        return NumberQuantity.of(1 / value, getUnit().inverse());
     }
 
     @Override
@@ -94,21 +94,23 @@ final class ByteQuantity<Q extends Quantity<Q>> extends JavaNumericQuantity<Q> {
 
     @Override
     public int getSize() {
-        return Byte.SIZE;
+        return Long.SIZE;
     }
 
     @Override
     public Class<?> getNumberType() {
-        return byte.class;
+        return long.class;
     }
 
     @Override
     Number castFromBigDecimal(BigDecimal aValue) {
-        return (byte) aValue.longValue();
+    	// FIXME https://github.com/unitsofmeasurement/indriya/issues/219 need to round properly, 13.99 must not turn to 13.
+        //return aValue.setScale(0, RoundingMode.HALF_UP).longValue(); would fix it but breaks LongQuantityTest.additionWithLargerMultipleAndOverflowingResultCastsToLargerMultiple
+    	return aValue.longValue();
     }
 
     @Override
     public Quantity<Q> negate() {
-        return new ByteQuantity<Q>((byte) (-value), getUnit());
+        return new LongQuantity<Q>(-value, getUnit());
     }
 }
