+++ src/test/java/tech/units/indriya/quantity/deprecated/ByteQuantityTest.java
@@ -27,7 +27,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tech.units.indriya.quantity;
+package tech.units.indriya.quantity.deprecated;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -48,23 +48,25 @@ import org.junit.jupiter.api.Test;
 
 import tech.units.indriya.AbstractUnit;
 import tech.units.indriya.quantity.Quantities;
-import tech.units.indriya.quantity.ShortQuantity;
+import tech.units.indriya.quantity.deprecated.ByteQuantity;
+import tech.units.indriya.quantity.deprecated.DoubleQuantity;
+import tech.units.indriya.quantity.deprecated.ShortQuantity;
 import tech.units.indriya.unit.Units;
 
-public class ShortQuantityTest {
+public class ByteQuantityTest {
 
+  private static final ShortQuantity<ElectricResistance> ONE_SHORT_OHM = new ShortQuantity<ElectricResistance>((short) 1, Units.OHM);
   private static final Unit<?> SQUARE_OHM = Units.OHM.multiply(Units.OHM);
-  private final ShortQuantity<ElectricResistance> ONE_OHM = createQuantity((short) 1, Units.OHM);
-  private final ShortQuantity<ElectricResistance> TWO_OHM = createQuantity((short) 2, Units.OHM);
-  private final ShortQuantity<ElectricResistance> MIN_VALUE_OHM = createQuantity(Short.MIN_VALUE, Units.OHM);
-  private final ShortQuantity<ElectricResistance> MAX_VALUE_OHM = createQuantity(Short.MAX_VALUE, Units.OHM);
-  private final ShortQuantity<ElectricResistance> ONE_MILLIOHM = createQuantity((short) 1, MetricPrefix.MILLI(Units.OHM));
-  private final ShortQuantity<ElectricResistance> ONE_KILOOHM = createQuantity((short) 1, MetricPrefix.KILO(Units.OHM));
-  private final ShortQuantity<ElectricResistance> ONE_YOTTAOHM = createQuantity((short) 1, MetricPrefix.YOTTA(Units.OHM));
-  private static final IntegerQuantity<ElectricResistance> ONE_INTEGER_OHM = new IntegerQuantity<ElectricResistance>(1, Units.OHM);
+  private final ByteQuantity<ElectricResistance> ONE_OHM = createQuantity((byte) 1, Units.OHM);
+  private final ByteQuantity<ElectricResistance> TWO_OHM = createQuantity((byte) 2, Units.OHM);
+  private final ByteQuantity<ElectricResistance> MIN_VALUE_OHM = createQuantity(Byte.MIN_VALUE, Units.OHM);
+  private final ByteQuantity<ElectricResistance> MAX_VALUE_OHM = createQuantity(Byte.MAX_VALUE, Units.OHM);
+  private final ByteQuantity<ElectricResistance> ONE_DEKAOHM = createQuantity((byte) 1, MetricPrefix.DEKA(Units.OHM));
+  private final ByteQuantity<ElectricResistance> ONE_DECIOHM = createQuantity((byte) 1, MetricPrefix.DECI(Units.OHM));
+  private final ByteQuantity<ElectricResistance> ONE_YOTTAOHM = createQuantity((byte) 1, MetricPrefix.YOTTA(Units.OHM));
 
-  private <Q extends Quantity<Q>> ShortQuantity<Q> createQuantity(short s, Unit<Q> unit) {
-    return new ShortQuantity<Q>(Short.valueOf(s).shortValue(), unit);
+  private static <Q extends Quantity<Q>> ByteQuantity<Q> createQuantity(byte b, Unit<Q> unit) {
+    return new ByteQuantity<Q>(Byte.valueOf(b).byteValue(), unit);
   }
 
   /**
@@ -74,7 +76,7 @@ public class ShortQuantityTest {
   @Test
   public void additionWithSameMultipleKeepsMultiple() {
     Quantity<ElectricResistance> actual = ONE_OHM.add(TWO_OHM);
-    ShortQuantity<ElectricResistance> expected = createQuantity((short) 3, Units.OHM);
+    ByteQuantity<ElectricResistance> expected = createQuantity((byte) 3, Units.OHM);
     assertEquals(expected, actual);
   }
 
@@ -93,8 +95,8 @@ public class ShortQuantityTest {
    */
   @Test
   public void additionWithLargerMultipleKeepsSmallerMultiple() {
-    Quantity<ElectricResistance> actual = ONE_MILLIOHM.add(ONE_OHM);
-    ShortQuantity<ElectricResistance> expected = createQuantity((short) 1001, MetricPrefix.MILLI(Units.OHM));
+    Quantity<ElectricResistance> actual = ONE_DECIOHM.add(ONE_OHM);
+    ByteQuantity<ElectricResistance> expected = createQuantity((byte) 11, MetricPrefix.DECI(Units.OHM));
     assertEquals(expected, actual);
   }
 
@@ -103,8 +105,8 @@ public class ShortQuantityTest {
    */
   @Test
   public void additionWithSmallerMultipleCastsToSmallerMultipleIfNeeded() {
-    Quantity<ElectricResistance> actual = ONE_OHM.add(ONE_MILLIOHM);
-    ShortQuantity<ElectricResistance> expected = createQuantity((short) 1001, MetricPrefix.MILLI(Units.OHM));
+    Quantity<ElectricResistance> actual = ONE_OHM.add(ONE_DECIOHM);
+    ByteQuantity<ElectricResistance> expected = createQuantity((byte) 11, MetricPrefix.DECI(Units.OHM));
     assertEquals(expected, actual);
   }
 
@@ -122,9 +124,9 @@ public class ShortQuantityTest {
    */
   @Test
   public void additionWithLargerMultipleAndOverflowingResultCastsToLargerMultiple() {
-    ShortQuantity<ElectricResistance> almost_max_value_ohm = createQuantity((short) (Short.MAX_VALUE - 999), Units.OHM);
-    Quantity<ElectricResistance> actual = almost_max_value_ohm.add(ONE_KILOOHM);
-    ShortQuantity<ElectricResistance> expected = createQuantity((short) (Short.MAX_VALUE / 1000), MetricPrefix.KILO(Units.OHM));
+    ByteQuantity<ElectricResistance> almost_max_value_ohm = createQuantity((byte) (Byte.MAX_VALUE - 9), Units.OHM);
+    Quantity<ElectricResistance> actual = almost_max_value_ohm.add(ONE_DEKAOHM);
+    ByteQuantity<ElectricResistance> expected = createQuantity((byte) (Byte.MAX_VALUE / 10), MetricPrefix.DEKA(Units.OHM));
     assertEquals(expected, actual);
   }
 
@@ -133,8 +135,8 @@ public class ShortQuantityTest {
    */
   @Test
   public void additionWithLargerMultipleButNotOverflowingResultKeepsSmallerMultiple() {
-    ShortQuantity<ElectricResistance> almost_max_value_ohm = createQuantity((short) (Short.MAX_VALUE - 1000), Units.OHM);
-    Quantity<ElectricResistance> actual = almost_max_value_ohm.add(ONE_KILOOHM);
+    ByteQuantity<ElectricResistance> almost_max_value_ohm = createQuantity((byte) (Byte.MAX_VALUE - 10), Units.OHM);
+    Quantity<ElectricResistance> actual = almost_max_value_ohm.add(ONE_DEKAOHM);
     assertEquals(MAX_VALUE_OHM, actual);
   }
 
@@ -171,7 +173,7 @@ public class ShortQuantityTest {
    */
   @Test
   public void subtractionWithSameMultipleAlmostResultingInNegativeDoesNotThrowException() {
-    Quantity<ElectricResistance> actual = createQuantity((short) (Short.MIN_VALUE + 1), Units.OHM).subtract(ONE_OHM);
+    Quantity<ElectricResistance> actual = createQuantity((byte) (Byte.MIN_VALUE + 1), Units.OHM).subtract(ONE_OHM);
     assertEquals(MIN_VALUE_OHM, actual);
   }
 
@@ -181,7 +183,7 @@ public class ShortQuantityTest {
   @Test
   public void quantityMultiplicationMultipliesCorrectly() {
     Quantity<?> actual = TWO_OHM.multiply(TWO_OHM);
-    ShortQuantity<?> expected = createQuantity((short) 4, SQUARE_OHM);
+    ByteQuantity<?> expected = createQuantity((byte) 4, SQUARE_OHM);
     assertEquals(expected, actual);
   }
 
@@ -191,7 +193,7 @@ public class ShortQuantityTest {
   @Test
   public void quantityMultiplicationResultingInOverflowThrowsException() {
     assertThrows(ArithmeticException.class, () -> {
-      Quantity<ElectricResistance> halfMaxValuePlusOne = createQuantity((short) (1 + Short.MAX_VALUE / 2), Units.OHM);
+      Quantity<ElectricResistance> halfMaxValuePlusOne = createQuantity((byte) (1 + Byte.MAX_VALUE / 2), Units.OHM);
       halfMaxValuePlusOne.multiply(TWO_OHM);
     });
   }
@@ -202,7 +204,7 @@ public class ShortQuantityTest {
   @Test
   public void numberMultiplicationMultipliesCorrectly() {
     Quantity<?> actual = TWO_OHM.multiply(2);
-    ShortQuantity<ElectricResistance> expected = createQuantity((short) 4, Units.OHM);
+    ByteQuantity<ElectricResistance> expected = createQuantity((byte) 4, Units.OHM);
     assertEquals(expected, actual);
   }
 
@@ -212,7 +214,7 @@ public class ShortQuantityTest {
   @Test
   public void numberMultiplicationResultingInOverflowThrowsException() {
     assertThrows(ArithmeticException.class, () -> {
-      Quantity<ElectricResistance> halfMaxValuePlusOne = createQuantity((short) (1 + Short.MAX_VALUE / 2), Units.OHM);
+      Quantity<ElectricResistance> halfMaxValuePlusOne = createQuantity((byte) (1 + Byte.MAX_VALUE / 2), Units.OHM);
       halfMaxValuePlusOne.multiply(2);
     });
   }
@@ -223,7 +225,7 @@ public class ShortQuantityTest {
   @Test
   public void quantityDivisionDividesCorrectly() {
     Quantity<?> actual = TWO_OHM.divide(TWO_OHM);
-    ShortQuantity<Dimensionless> expected = createQuantity((short) 1, AbstractUnit.ONE);
+    ByteQuantity<Dimensionless> expected = createQuantity((byte) 1, AbstractUnit.ONE);
     assertEquals(expected, actual);
   }
 
@@ -242,7 +244,7 @@ public class ShortQuantityTest {
   @Test
   public void inverseReturnsUnitQuantityForUnitQuantity() {
     Quantity<?> actual = ONE_OHM.inverse();
-    ShortQuantity<?> expected = createQuantity((short) 1, Units.OHM.inverse());
+    ByteQuantity<?> expected = createQuantity((byte) 1, Units.OHM.inverse());
     assertEquals(expected, actual);
   }
 
@@ -252,7 +254,7 @@ public class ShortQuantityTest {
   @Test
   public void inverseReturnsZeroQuantityForLargerThanUnitQuantity() {
     Quantity<?> actual = TWO_OHM.inverse();
-    ShortQuantity<?> expected = createQuantity((short) 0, Units.OHM.inverse());
+    ByteQuantity<?> expected = createQuantity((byte) 0, Units.OHM.inverse());
     assertEquals(expected, actual);
   }
 
@@ -262,39 +264,39 @@ public class ShortQuantityTest {
   @Test
   public void inverseThrowsExceptionForZeroQuantity() {
     assertThrows(ArithmeticException.class, () -> {
-      createQuantity((short) 0, Units.OHM).inverse();
+      createQuantity((byte) 0, Units.OHM).inverse();
     });
   }
 
   /**
-   * Verifies that a ShortQuantity isn't big.
+   * Verifies that a ByteQuantity isn't big.
    */
   @Test
-  public void shortQuantityIsNotBig() {
+  public void byteQuantityIsNotBig() {
     assertFalse(ONE_OHM.isBig());
   }
 
   /**
-   * Verifies that a ShortQuantity isn't decimal.
+   * Verifies that a ByteQuantity isn't decimal.
    */
   @Test
-  public void shortQuantityIsNotDecimal() {
+  public void byteQuantityIsNotDecimal() {
     assertFalse(ONE_OHM.isDecimal());
   }
 
   /**
-   * Verifies that a ShortQuantity has the size of Short.
+   * Verifies that a ByteQuantity has the size of Byte.
    */
   @Test
-  public void shortQuantityHasByteSize() {
-    assertEquals(Short.SIZE, ONE_OHM.getSize());
+  public void byteQuantityHasByteSize() {
+    assertEquals(Byte.SIZE, ONE_OHM.getSize());
   }
 
   /**
    * Verifies that a quantity isn't equal to null.
    */
   @Test
-  public void shortQuantityIsNotEqualToNull() {
+  public void byteQuantityIsNotEqualToNull() {
     assertFalse(ONE_OHM.equals(null));
   }
 
@@ -302,7 +304,7 @@ public class ShortQuantityTest {
    * Verifies that a quantity is equal to itself.
    */
   @Test
-  public void shortQuantityIsEqualToItself() {
+  public void byteQuantityIsEqualToItself() {
     assertTrue(ONE_OHM.equals(ONE_OHM));
   }
 
@@ -310,15 +312,15 @@ public class ShortQuantityTest {
    * Verifies that a quantity is equal to another instance with the same value and unit.
    */
   @Test
-  public void shortQuantityIsEqualToIdenticalInstance() {
-    assertTrue(ONE_OHM.equals(createQuantity((short) 1, Units.OHM)));
+  public void byteQuantityIsEqualToIdenticalInstance() {
+    assertTrue(ONE_OHM.equals(createQuantity((byte) 1, Units.OHM)));
   }
 
   /**
    * Verifies that a quantity is equal to another instance with the same value and unit using another primitive.
    */
   @Test
-  public void shortQuantityIsEqualToIdenticalInstanceWithAnotherPrimitive() {
+  public void byteQuantityIsEqualToIdenticalInstanceWithAnotherPrimitive() {
     assertTrue(ONE_OHM.equals(new DoubleQuantity<ElectricResistance>(Double.valueOf(1).doubleValue(), Units.OHM)));
   }
 
@@ -326,7 +328,7 @@ public class ShortQuantityTest {
    * Verifies that a quantity is not equal to a quantity with a different value.
    */
   @Test
-  public void shortQuantityIsNotEqualToQuantityWithDifferentValue() {
+  public void byteQuantityIsNotEqualToQuantityWithDifferentValue() {
     assertFalse(ONE_OHM.equals(TWO_OHM));
   }
 
@@ -334,15 +336,15 @@ public class ShortQuantityTest {
    * Verifies that a quantity is not equal to a quantity with a different unit.
    */
   @Test
-  public void shortQuantityIsNotEqualToQuantityWithDifferentUnit() {
-    assertFalse(ONE_OHM.equals(ONE_KILOOHM));
+  public void byteQuantityIsNotEqualToQuantityWithDifferentUnit() {
+    assertFalse(ONE_OHM.equals(ONE_DEKAOHM));
   }
 
   /**
    * Verifies that a quantity is not equal to an object of a different class.
    */
   @Test
-  public void shortQuantityIsNotEqualToObjectOfDifferentClass() {
+  public void byteQuantityIsNotEqualToObjectOfDifferentClass() {
     assertFalse(ONE_OHM.equals(SQUARE_OHM));
   }
 
@@ -359,7 +361,7 @@ public class ShortQuantityTest {
    */
   @Test
   public void doubleValueReturnsConvertedValueForOtherUnit() {
-    assertEquals(0.001, ONE_MILLIOHM.doubleValue(Units.OHM));
+    assertEquals(0.1, ONE_DECIOHM.doubleValue(Units.OHM));
   }
 
   /**
@@ -375,7 +377,7 @@ public class ShortQuantityTest {
    */
   @Test
   public void decimalValueReturnsConvertedValueForOtherUnit() {
-    assertEquals(BigDecimal.valueOf(0.001), ONE_MILLIOHM.decimalValue(Units.OHM));
+    assertEquals(BigDecimal.valueOf(0.1), ONE_DECIOHM.decimalValue(Units.OHM));
   }
 
   /**
@@ -391,7 +393,7 @@ public class ShortQuantityTest {
    */
   @Test
   public void longValueReturnsConvertedValueForOtherUnit() {
-    assertEquals(0, ONE_MILLIOHM.longValue(Units.OHM));
+    assertEquals(0, ONE_DECIOHM.longValue(Units.OHM));
   }
 
   /**
@@ -400,7 +402,7 @@ public class ShortQuantityTest {
   @Test
   public void longValueThrowsExceptionOnPositiveOverflow() {
     assertThrows(ArithmeticException.class, () -> {
-      createQuantity((short) 9224, MetricPrefix.PETA(Units.OHM)).longValue(Units.OHM);
+      createQuantity((byte) 10, MetricPrefix.EXA(Units.OHM)).longValue(Units.OHM);
     });
   }
 
@@ -410,72 +412,72 @@ public class ShortQuantityTest {
   @Test
   public void longValueThrowsExceptionOnNegativeOverflow() {
     assertThrows(ArithmeticException.class, () -> {
-      createQuantity((short) -9224, MetricPrefix.PETA(Units.OHM)).longValue(Units.OHM);
+      createQuantity((byte) -10, MetricPrefix.EXA(Units.OHM)).longValue(Units.OHM);
     });
   }
 
   /**
-   * Verifies that addition with ShortQuantity returns a ShortQuantity.
+   * Verifies that addition with ByteQuantity returns a ByteQuantity.
    */
   @Test
-  public void additionWithShortQuantityDoesNotWiden() {
-    assertEquals(ShortQuantity.class, ONE_OHM.add(ONE_OHM).getClass());
+  public void additionWithByteQuantityDoesNotWiden() {
+    assertEquals(ByteQuantity.class, ONE_OHM.add(ONE_OHM).getClass());
   }
 
   /**
-   * Verifies that addition with IntegerQuantity widens to IntegerQuantity.
+   * Verifies that addition with ShortQuantity widens to ShortQuantity.
    */
   @Test
-  public void additionWithIntegerQuantityWidensToIntegerQuantity() {
-    assertEquals(IntegerQuantity.class, ONE_OHM.add(ONE_INTEGER_OHM).getClass());
+  public void additionWithShortQuantityWidensToShortQuantity() {
+    assertEquals(ShortQuantity.class, ONE_OHM.add(ONE_SHORT_OHM).getClass());
   }
 
   /**
-   * Verifies that subtraction with ShortQuantity returns a ShortQuantity.
+   * Verifies that subtraction with ByteQuantity returns a ByteQuantity.
    */
   @Test
-  public void subtractionWithShortQuantityDoesNotWiden() {
-    assertEquals(ShortQuantity.class, ONE_OHM.subtract(ONE_OHM).getClass());
+  public void subtractionWithByteQuantityDoesNotWiden() {
+    assertEquals(ByteQuantity.class, ONE_OHM.subtract(ONE_OHM).getClass());
   }
 
   /**
-   * Verifies that subtraction with IntegerQuantity widens to IntegerQuantity.
+   * Verifies that subtraction with ShortQuantity widens to ShortQuantity.
    */
   @Test
-  public void subtractionWithIntegerQuantityWidensToIntegerQuantity() {
-    assertEquals(IntegerQuantity.class, ONE_OHM.subtract(ONE_INTEGER_OHM).getClass());
+  public void subtractionWithShortQuantityWidensToShortQuantity() {
+    assertEquals(ShortQuantity.class, ONE_OHM.subtract(ONE_SHORT_OHM).getClass());
   }
 
   /**
-   * Verifies that multiplication with ShortQuantity returns a ShortQuantity.
+   * Verifies that multiplication with ByteQuantity returns a ByteQuantity.
    */
   @Test
-  public void multiplicationWithShortQuantityDoesNotWiden() {
-    assertEquals(ShortQuantity.class, ONE_OHM.multiply(ONE_OHM).getClass());
+  public void multiplicationWithByteQuantityDoesNotWiden() {
+    assertEquals(ByteQuantity.class, ONE_OHM.multiply(ONE_OHM).getClass());
   }
 
   /**
-   * Verifies that multiplication with IntegerQuantity widens to IntegerQuantity.
+   * Verifies that multiplication with ShortQuantity widens to ShortQuantity.
    */
   @Test
-  public void multiplicationWithIntegerQuantityWidensToIntegerQuantity() {
-    assertEquals(IntegerQuantity.class, ONE_OHM.multiply(ONE_INTEGER_OHM).getClass());
+  public void multiplicationWithShortQuantityWidensToShortQuantity() {
+    assertEquals(ShortQuantity.class, ONE_OHM.multiply(ONE_SHORT_OHM).getClass());
   }
 
   /**
-   * Verifies that division with ShortQuantity returns a ShortQuantity.
+   * Verifies that division with ByteQuantity returns a ByteQuantity.
    */
   @Test
-  public void divisionWithShortQuantityDoesNotWiden() {
-    assertEquals(ShortQuantity.class, ONE_OHM.divide(ONE_OHM).getClass());
+  public void divisionWithByteQuantityDoesNotWiden() {
+    assertEquals(ByteQuantity.class, ONE_OHM.divide(ONE_OHM).getClass());
   }
 
   /**
-   * Verifies that division with IntegerQuantity widens to IntegerQuantity.
+   * Verifies that division with ShortQuantity widens to ShortQuantity.
    */
   @Test
-  public void divisionWithIntegerQuantityWidensToIntegerQuantity() {
-    assertEquals(IntegerQuantity.class, ONE_OHM.divide(ONE_INTEGER_OHM).getClass());
+  public void divisionWithShortQuantityWidensToShortQuantity() {
+    assertEquals(ShortQuantity.class, ONE_OHM.divide(ONE_SHORT_OHM).getClass());
   }
 
   @Test
@@ -490,18 +492,18 @@ public class ShortQuantityTest {
     assertEquals(dayResult.getValue().intValue(), day.getValue().intValue());
   }
 
-  @Test
-  public void testEquality() throws Exception {
-    Quantity<Length> value = Quantities.getQuantity(Short.valueOf("1"), Units.METRE);
-    Quantity<Length> anotherValue = Quantities.getQuantity(Short.valueOf("1"), Units.METRE);
-    assertEquals(value, anotherValue);
-  }
-
   /**
    * Tests negate()
    */
   @Test
   public void negateTest() {
-    assertEquals((short) -1, ONE_OHM.negate().getValue());
+    assertEquals((byte) -1, ONE_OHM.negate().getValue());
+  }
+
+  @Test
+  public void testEquality() throws Exception {
+    Quantity<Length> value = Quantities.getQuantity(Byte.valueOf("1"), Units.METRE);
+    Quantity<Length> anotherValue = Quantities.getQuantity(Byte.valueOf("1"), Units.METRE);
+    assertEquals(value, anotherValue);
   }
 }
