+++ src/test/java/tech/units/indriya/quantity/deprecated/DecimalQuantityTest.java
@@ -27,7 +27,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tech.units.indriya.quantity;
+package tech.units.indriya.quantity.deprecated;
 
 import static javax.measure.MetricPrefix.MILLI;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -48,6 +48,9 @@ import javax.measure.quantity.Time;
 
 import org.junit.jupiter.api.Test;
 
+import tech.units.indriya.quantity.Quantities;
+import tech.units.indriya.quantity.deprecated.DecimalQuantity;
+import tech.units.indriya.quantity.deprecated.DoubleQuantity;
 import tech.units.indriya.unit.Units;
 
 public class DecimalQuantityTest {
