+++ src/test/java/tech/units/indriya/quantity/deprecated/DoubleQuantityTest.java
@@ -27,7 +27,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tech.units.indriya.quantity;
+package tech.units.indriya.quantity.deprecated;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -36,7 +36,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.math.BigDecimal;
 
-import javax.measure.MetricPrefix;
 import javax.measure.Quantity;
 import javax.measure.Unit;
 import javax.measure.quantity.Dimensionless;
@@ -46,24 +45,28 @@ import javax.measure.quantity.Time;
 import org.junit.jupiter.api.Test;
 
 import tech.units.indriya.AbstractUnit;
-import tech.units.indriya.quantity.FloatQuantity;
 import tech.units.indriya.quantity.Quantities;
+import tech.units.indriya.quantity.deprecated.DecimalQuantity;
+import tech.units.indriya.quantity.deprecated.DoubleQuantity;
+import tech.units.indriya.quantity.deprecated.LongQuantity;
+
 import static javax.measure.MetricPrefix.*;
 import tech.units.indriya.unit.Units;
 
-public class FloatQuantityTest {
+public class DoubleQuantityTest {
 
   private static final Unit<?> SQUARE_OHM = Units.OHM.multiply(Units.OHM);
-  private static final FloatQuantity<ElectricResistance> HALF_AN_OHM = createQuantity(0.5F, Units.OHM);
-  private static final FloatQuantity<ElectricResistance> ONE_OHM = createQuantity(1, Units.OHM);
-  private static final FloatQuantity<ElectricResistance> TWO_OHM = createQuantity(2, Units.OHM);
-  private static final FloatQuantity<ElectricResistance> MAX_VALUE_OHM = createQuantity(Float.MAX_VALUE, Units.OHM);
-  private static final FloatQuantity<ElectricResistance> JUST_OVER_HALF_MAX_VALUE_OHM = createQuantity(Float.MAX_VALUE / 1.9999999F, Units.OHM);
-  private static final FloatQuantity<ElectricResistance> ONE_MILLIOHM = createQuantity(1, MILLI(Units.OHM));
-  private static final DoubleQuantity<ElectricResistance> ONE_DOUBLE_OHM = new DoubleQuantity<ElectricResistance>(1D, Units.OHM);
+  private static final DoubleQuantity<ElectricResistance> HALF_AN_OHM = createQuantity(0.5, Units.OHM);
+  private static final DoubleQuantity<ElectricResistance> ONE_OHM = createQuantity(1, Units.OHM);
+  private static final DoubleQuantity<ElectricResistance> TWO_OHM = createQuantity(2, Units.OHM);
+  private static final DoubleQuantity<ElectricResistance> MAX_VALUE_OHM = createQuantity(Double.MAX_VALUE, Units.OHM);
+  private static final DoubleQuantity<ElectricResistance> JUST_OVER_HALF_MAX_VALUE_OHM = createQuantity(Double.MAX_VALUE / 1.999999999999999,
+      Units.OHM);
+  private static final DoubleQuantity<ElectricResistance> ONE_MILLIOHM = createQuantity(1, MILLI(Units.OHM));
+  private static final DecimalQuantity<ElectricResistance> ONE_DECIMAL_OHM = new DecimalQuantity<ElectricResistance>(BigDecimal.ONE, Units.OHM);
 
-  private static <Q extends Quantity<Q>> FloatQuantity<Q> createQuantity(float f, Unit<Q> unit) {
-    return new FloatQuantity<Q>(Float.valueOf(f).floatValue(), unit);
+  private static <Q extends Quantity<Q>> DoubleQuantity<Q> createQuantity(double d, Unit<Q> unit) {
+    return new DoubleQuantity<Q>(Double.valueOf(d).doubleValue(), unit);
   }
 
   /**
@@ -73,7 +76,7 @@ public class FloatQuantityTest {
   @Test
   public void additionWithSameMultipleKeepsMultiple() {
     Quantity<ElectricResistance> actual = ONE_OHM.add(TWO_OHM);
-    FloatQuantity<ElectricResistance> expected = createQuantity(3, Units.OHM);
+    DoubleQuantity<ElectricResistance> expected = createQuantity(3, Units.OHM);
     assertEquals(expected, actual);
   }
 
@@ -83,7 +86,7 @@ public class FloatQuantityTest {
   @Test
   public void additionWithLargerMultipleKeepsSmallerMultiple() {
     Quantity<ElectricResistance> actual = ONE_MILLIOHM.add(ONE_OHM);
-    FloatQuantity<ElectricResistance> expected = createQuantity(1001, MetricPrefix.MILLI(Units.OHM));
+    DoubleQuantity<ElectricResistance> expected = createQuantity(1001, MILLI(Units.OHM));
     assertEquals(expected, actual);
   }
 
@@ -93,7 +96,7 @@ public class FloatQuantityTest {
   @Test
   public void additionWithSmallerMultipleKeepsLargerMultiple() {
     Quantity<ElectricResistance> actual = ONE_OHM.add(ONE_MILLIOHM);
-    FloatQuantity<ElectricResistance> expected = createQuantity(1.001f, Units.OHM);
+    DoubleQuantity<ElectricResistance> expected = createQuantity(1.001, Units.OHM);
     assertEquals(expected, actual);
   }
 
@@ -112,8 +115,9 @@ public class FloatQuantityTest {
    */
   @Test
   public void additionWithLargerOverflowingMultipleCastsToLargerMultiple() {
-    Quantity<ElectricResistance> actual = ONE_MILLIOHM.add(MAX_VALUE_OHM);
-    assertEquals(MAX_VALUE_OHM, actual);
+    Quantity<ElectricResistance> largeValueOhm = MAX_VALUE_OHM.divide(2);
+    Quantity<ElectricResistance> actual = ONE_MILLIOHM.add(largeValueOhm);
+    assertEquals(largeValueOhm, actual);
   }
 
   /**
@@ -123,6 +127,195 @@ public class FloatQuantityTest {
   public void subtractionSubtractsArgumentFromTargetObject() {
     Quantity<ElectricResistance> actual = TWO_OHM.subtract(ONE_OHM);
     assertEquals(ONE_OHM, actual);
+  }
+
+  /**
+   * Verifies that the multiplication of two quantities multiplies correctly.
+   */
+  @Test
+  public void quantityMultiplicationMultipliesCorrectly() {
+    Quantity<?> actual = TWO_OHM.multiply(TWO_OHM);
+    DoubleQuantity<?> expected = createQuantity(4, SQUARE_OHM);
+    assertEquals(expected, actual);
+  }
+
+  /**
+   * Verifies that the multiplication of two quantities resulting in an overflow throws an exception.
+   */
+  @Test
+  public void quantityMultiplicationResultingInOverflowThrowsException() {
+    assertThrows(ArithmeticException.class, () -> {
+      JUST_OVER_HALF_MAX_VALUE_OHM.multiply(TWO_OHM);
+    });
+  }
+
+  /**
+   * Verifies that the multiplication with a number multiplies correctly.
+   */
+  @Test
+  public void numberMultiplicationMultipliesCorrectly() {
+    Quantity<?> actual = TWO_OHM.multiply(2);
+    DoubleQuantity<ElectricResistance> expected = createQuantity(4, Units.OHM);
+    assertEquals(expected, actual);
+  }
+
+  /**
+   * Verifies that the multiplication with a number resulting in an overflow throws an exception.
+   */
+  @Test
+  public void numberMultiplicationResultingInOverflowThrowsException() {
+    assertThrows(ArithmeticException.class, () -> {
+      JUST_OVER_HALF_MAX_VALUE_OHM.multiply(2);
+    });
+  }
+
+  /**
+   * Verifies that the division of two quantities divides correctly.
+   */
+  @Test
+  public void quantityDivisionDividesCorrectly() {
+    Quantity<?> actual = TWO_OHM.divide(TWO_OHM);
+    DoubleQuantity<Dimensionless> expected = createQuantity(1, AbstractUnit.ONE);
+    assertEquals(expected, actual);
+  }
+
+  /**
+   * Verifies that the division of two quantities resulting in an overflow throws an exception.
+   */
+  @Test
+  public void quantityDivisionResultingInOverflowThrowsException() {
+    assertThrows(ArithmeticException.class, () -> {
+      JUST_OVER_HALF_MAX_VALUE_OHM.divide(HALF_AN_OHM);
+    });
+  }
+
+  /**
+   * Verifies that the division with a number divides correctly.
+   */
+  @Test
+  public void numberDivisionDividesCorrectly() {
+    Quantity<?> actual = TWO_OHM.divide(2);
+    assertEquals(ONE_OHM, actual);
+  }
+
+  /**
+   * Verifies that the division with a number resulting in an overflow throws an exception.
+   */
+  @Test
+  public void numberDivisionResultingInOverflowThrowsException() {
+    assertThrows(ArithmeticException.class, () -> {
+      JUST_OVER_HALF_MAX_VALUE_OHM.divide(0.5);
+    });
+  }
+
+  /**
+   * Verifies that the inverse returns the correct reciprocal for a unit quantity.
+   */
+  @Test
+  public void inverseReturnsUnitQuantityForUnitQuantity() {
+    Quantity<?> actual = ONE_OHM.inverse();
+    DoubleQuantity<?> expected = createQuantity(1, Units.OHM.inverse());
+    assertEquals(expected, actual);
+  }
+
+  /**
+   * Verifies that the inverse returns the correct reciprocal for a quantity larger than a unit quantity.
+   */
+  @Test
+  public void inverseReturnsZeroQuantityForLargerThanUnitQuantity() {
+    Quantity<?> actual = TWO_OHM.inverse();
+    DoubleQuantity<?> expected = createQuantity(0.5, Units.OHM.inverse());
+    assertEquals(expected, actual);
+  }
+
+  /**
+   * Verifies that the inverse throws an exception for a zero quantity.
+   */
+  @Test
+  public void inverseReturnsInfinityQuantityForZeroQuantity() {
+    Quantity<?> actual = createQuantity(0, Units.OHM).inverse();
+    DoubleQuantity<?> expected = createQuantity(Double.POSITIVE_INFINITY, Units.OHM.inverse());
+    assertEquals(expected, actual);
+  }
+
+  /**
+   * Verifies that a DoubleQuantity isn't big.
+   */
+  @Test
+  public void doubleQuantityIsNotBig() {
+    assertFalse(ONE_OHM.isBig());
+  }
+
+  /**
+   * Verifies that a DoubleQuantity is decimal.
+   */
+  @Test
+  public void doubleQuantityIsDecimal() {
+    assertTrue(ONE_OHM.isDecimal());
+  }
+
+  /**
+   * Verifies that a DoubleQuantity has the size of Double.
+   */
+  @Test
+  public void doubleQuantityHasByteSize() {
+    assertEquals(Double.SIZE, ONE_OHM.getSize());
+  }
+
+  /**
+   * Verifies that a quantity isn't equal to null.
+   */
+  @Test
+  public void doubleQuantityIsNotEqualToNull() {
+    assertFalse(ONE_OHM.equals(null));
+  }
+
+  /**
+   * Verifies that a quantity is equal to itself.
+   */
+  @Test
+  public void doubleQuantityIsEqualToItself() {
+    assertTrue(ONE_OHM.equals(ONE_OHM));
+  }
+
+  /**
+   * Verifies that a quantity is equal to another instance with the same value and unit.
+   */
+  @Test
+  public void doubleQuantityIsEqualToIdenticalInstance() {
+    assertTrue(ONE_OHM.equals(createQuantity(1, Units.OHM)));
+  }
+
+  /**
+   * Verifies that a quantity is equal to another instance with the same value and unit using another primitive.
+   */
+  @Test
+  public void doubleQuantityIsEqualToIdenticalInstanceWithAnotherPrimitive() {
+    assertTrue(ONE_OHM.equals(new LongQuantity<ElectricResistance>(Long.valueOf(1L).longValue(), Units.OHM)));
+  }
+
+  /**
+   * Verifies that a quantity is not equal to a quantity with a different value.
+   */
+  @Test
+  public void doubleQuantityIsNotEqualToQuantityWithDifferentValue() {
+    assertFalse(ONE_OHM.equals(TWO_OHM));
+  }
+
+  /**
+   * Verifies that a quantity is not equal to a quantity with a different unit.
+   */
+  @Test
+  public void doubleQuantityIsNotEqualToQuantityWithDifferentUnit() {
+    assertFalse(ONE_OHM.equals(ONE_MILLIOHM));
+  }
+
+  /**
+   * Verifies that a quantity is not equal to an object of a different class.
+   */
+  @Test
+  public void doubleQuantityIsNotEqualToObjectOfDifferentClass() {
+    assertFalse(ONE_OHM.equals(SQUARE_OHM));
   }
 
   /**
@@ -139,6 +332,16 @@ public class FloatQuantityTest {
   @Test
   public void doubleValueReturnsConvertedValueForOtherUnit() {
     assertEquals(0.001, ONE_MILLIOHM.doubleValue(Units.OHM));
+  }
+
+  /**
+   * Verifies that an exception is thrown if the conversion for doubleValue results in overflow.
+   */
+  @Test
+  public void doubleValueThrowsExceptionOnOverflow() {
+    assertThrows(ArithmeticException.class, () -> {
+      MAX_VALUE_OHM.doubleValue(MILLI(Units.OHM));
+    });
   }
 
   /**
@@ -179,7 +382,7 @@ public class FloatQuantityTest {
   @Test
   public void longValueThrowsExceptionOnPositiveOverflow() {
     assertThrows(ArithmeticException.class, () -> {
-      createQuantity(Long.MAX_VALUE + 550000000000.0F, Units.OHM).longValue(Units.OHM);
+      createQuantity(Long.MAX_VALUE + 1025.0, Units.OHM).longValue(Units.OHM);
     });
   }
 
@@ -189,99 +392,84 @@ public class FloatQuantityTest {
   @Test
   public void longValueThrowsExceptionOnNegativeOverflow() {
     assertThrows(ArithmeticException.class, () -> {
-      createQuantity(Long.MIN_VALUE - 550000000000.0F, Units.OHM).longValue(Units.OHM);
+      createQuantity(Long.MIN_VALUE - 1025.0, Units.OHM).longValue(Units.OHM);
     });
   }
 
   /**
-   * Verifies that the multiplication of two quantities multiplies correctly.
+   * Verifies that addition with DoubleQuantity returns a DoubleQuantity.
    */
   @Test
-  public void quantityMultiplicationMultipliesCorrectly() {
-    Quantity<?> actual = TWO_OHM.multiply(TWO_OHM);
-    FloatQuantity<?> expected = createQuantity(4, SQUARE_OHM);
-    assertEquals(expected, actual);
+  public void additionWithDoubleQuantityDoesNotWiden() {
+    assertEquals(DoubleQuantity.class, ONE_OHM.add(ONE_OHM).getClass());
   }
 
   /**
-   * Verifies that the multiplication of two quantities resulting in an overflow throws an exception.
+   * Verifies that addition with DecimalQuantity widens to DecimalQuantity.
    */
   @Test
-  public void quantityMultiplicationResultingInOverflowThrowsException() {
-    assertThrows(ArithmeticException.class, () -> {
-      JUST_OVER_HALF_MAX_VALUE_OHM.multiply(TWO_OHM);
-    });
+  public void additionWithDecimalQuantityWidensToDecimalQuantity() {
+    assertEquals(DecimalQuantity.class, ONE_OHM.add(ONE_DECIMAL_OHM).getClass());
   }
 
   /**
-   * Verifies that the multiplication with a number multiplies correctly.
+   * Verifies that subtraction with DoubleQuantity returns a DoubleQuantity.
    */
   @Test
-  public void numberMultiplicationMultipliesCorrectly() {
-    Quantity<?> actual = TWO_OHM.multiply(2);
-    FloatQuantity<ElectricResistance> expected = createQuantity(4, Units.OHM);
-    assertEquals(expected, actual);
+  public void subtractionWithDoubleQuantityDoesNotWiden() {
+    assertEquals(DoubleQuantity.class, ONE_OHM.subtract(ONE_OHM).getClass());
   }
 
   /**
-   * Verifies that the multiplication with a number resulting in an overflow throws an exception.
+   * Verifies that subtraction with DecimalQuantity widens to DecimalQuantity.
    */
   @Test
-  public void numberMultiplicationResultingInOverflowThrowsException() {
-    assertThrows(ArithmeticException.class, () -> {
-      JUST_OVER_HALF_MAX_VALUE_OHM.multiply(2F);
-    });
+  public void subtractionWithDecimalQuantityWidensToDecimalQuantity() {
+    assertEquals(DecimalQuantity.class, ONE_OHM.subtract(ONE_DECIMAL_OHM).getClass());
   }
 
   /**
-   * Verifies that the division of two quantities divides correctly.
+   * Verifies that multiplication with DoubleQuantity returns a DoubleQuantity.
    */
   @Test
-  public void quantityDivisionDividesCorrectly() {
-    Quantity<?> actual = TWO_OHM.divide(TWO_OHM);
-    FloatQuantity<Dimensionless> expected = createQuantity(1, AbstractUnit.ONE);
-    assertEquals(expected, actual);
+  public void multiplicationWithDoubleQuantityDoesNotWiden() {
+    assertEquals(DoubleQuantity.class, ONE_OHM.multiply(ONE_OHM).getClass());
   }
 
   /**
-   * Verifies that the division of two quantities resulting in an overflow throws an exception.
+   * Verifies that multiplication with DecimalQuantity widens to DecimalQuantity.
    */
   @Test
-  public void quantityDivisionResultingInOverflowThrowsException() {
-    assertThrows(ArithmeticException.class, () -> {
-      JUST_OVER_HALF_MAX_VALUE_OHM.divide(HALF_AN_OHM);
-    });
+  public void multiplicationWithDecimalQuantityWidensToDecimalQuantity() {
+    assertEquals(DecimalQuantity.class, ONE_OHM.multiply(ONE_DECIMAL_OHM).getClass());
   }
 
   /**
-   * Verifies that the division with a number divides correctly.
+   * Verifies that division with DoubleQuantity returns a DoubleQuantity.
    */
   @Test
-  public void numberDivisionDividesCorrectly() {
-    Quantity<?> actual = TWO_OHM.divide(2);
-    assertEquals(ONE_OHM, actual);
+  public void divisionWithDoubleQuantityDoesNotWiden() {
+    assertEquals(DoubleQuantity.class, ONE_OHM.divide(ONE_OHM).getClass());
   }
 
   /**
-   * Verifies that the division with a number resulting in an overflow throws an exception.
+   * Verifies that division with DecimalQuantity widens to DecimalQuantity.
    */
   @Test
-  public void numberDivisionResultingInOverflowThrowsException() {
-    assertThrows(ArithmeticException.class, () -> {
-      JUST_OVER_HALF_MAX_VALUE_OHM.divide(0.5F);
-    });
+  public void divisionWithDecimalQuantityWidensToDecimalQuantity() {
+    assertEquals(DecimalQuantity.class, ONE_OHM.divide(ONE_DECIMAL_OHM).getClass());
   }
 
   @Test
   public void toTest() {
     Quantity<Time> day = Quantities.getQuantity(1D, Units.DAY);
     Quantity<Time> hour = day.to(Units.HOUR);
-    assertEquals(Double.valueOf(24), hour.getValue());
+    assertEquals(hour.getValue().intValue(), 24);
     assertEquals(hour.getUnit(), Units.HOUR);
 
     Quantity<Time> dayResult = hour.to(Units.DAY);
-    assertEquals(dayResult.getValue(), day.getValue());
-    assertEquals(dayResult.getUnit(), day.getUnit());
+    assertEquals(dayResult.getValue().intValue(), day.getValue().intValue());
+    assertEquals(dayResult.getValue().intValue(), day.getValue().intValue());
   }
 
   /**
@@ -289,180 +477,6 @@ public class FloatQuantityTest {
    */
   @Test
   public void negateTest() {
-    assertEquals(-1f, ONE_OHM.negate().getValue());
-  }
-
-  /**
-   * Verifies that the inverse returns the correct reciprocal for a unit quantity.
-   */
-  @Test
-  public void inverseReturnsUnitQuantityForUnitQuantity() {
-    Quantity<?> actual = ONE_OHM.inverse();
-    FloatQuantity<?> expected = createQuantity(1, Units.OHM.inverse());
-    assertEquals(expected, actual);
-  }
-
-  /**
-   * Verifies that the inverse returns the correct reciprocal for a quantity larger than a unit quantity.
-   */
-  @Test
-  public void inverseReturnsZeroQuantityForLargerThanUnitQuantity() {
-    Quantity<?> actual = TWO_OHM.inverse();
-    FloatQuantity<?> expected = createQuantity(0.5F, Units.OHM.inverse());
-    assertEquals(expected, actual);
-  }
-
-  /**
-   * Verifies that the inverse throws an exception for a zero quantity.
-   */
-  @Test
-  public void inverseReturnsInfinityQuantityForZeroQuantity() {
-    Quantity<?> actual = createQuantity(0, Units.OHM).inverse();
-    FloatQuantity<?> expected = createQuantity(Float.POSITIVE_INFINITY, Units.OHM.inverse());
-    assertEquals(expected, actual);
-  }
-
-  /**
-   * Verifies that a FloatQuantity isn't big.
-   */
-  @Test
-  public void floatQuantityIsNotBig() {
-    assertFalse(ONE_OHM.isBig());
-  }
-
-  /**
-   * Verifies that a FloatQuantity is decimal.
-   */
-  @Test
-  public void floatQuantityIsDecimal() {
-    assertTrue(ONE_OHM.isDecimal());
-  }
-
-  /**
-   * Verifies that a FloatQuantity has the size of Float.
-   */
-  @Test
-  public void floatQuantityHasByteSize() {
-    assertEquals(Float.SIZE, ONE_OHM.getSize());
-  }
-
-  /**
-   * Verifies that a quantity isn't equal to null.
-   */
-  @Test
-  public void floatQuantityIsNotEqualToNull() {
-    assertFalse(ONE_OHM.equals(null));
-  }
-
-  /**
-   * Verifies that a quantity is equal to itself.
-   */
-  @Test
-  public void floatQuantityIsEqualToItself() {
-    assertTrue(ONE_OHM.equals(ONE_OHM));
-  }
-
-  /**
-   * Verifies that a quantity is equal to another instance with the same value and unit.
-   */
-  @Test
-  public void floatQuantityIsEqualToIdenticalInstance() {
-    assertTrue(ONE_OHM.equals(createQuantity(1, Units.OHM)));
-  }
-
-  /**
-   * Verifies that a quantity is equal to another instance with the same value and unit using another primitive.
-   */
-  @Test
-  public void floatQuantityIsEqualToIdenticalInstanceWithAnotherPrimitive() {
-    assertTrue(ONE_OHM.equals(new LongQuantity<ElectricResistance>(Long.valueOf(1L).longValue(), Units.OHM)));
-  }
-
-  /**
-   * Verifies that a quantity is not equal to a quantity with a different value.
-   */
-  @Test
-  public void floatQuantityIsNotEqualToQuantityWithDifferentValue() {
-    assertFalse(ONE_OHM.equals(TWO_OHM));
-  }
-
-  /**
-   * Verifies that a quantity is not equal to a quantity with a different unit.
-   */
-  @Test
-  public void floatQuantityIsNotEqualToQuantityWithDifferentUnit() {
-    assertFalse(ONE_OHM.equals(ONE_MILLIOHM));
-  }
-
-  /**
-   * Verifies that a quantity is not equal to an object of a different class.
-   */
-  @Test
-  public void floatQuantityIsNotEqualToObjectOfDifferentClass() {
-    assertFalse(ONE_OHM.equals(SQUARE_OHM));
-  }
-
-  /**
-   * Verifies that addition with FloatQuantity returns a FloatQuantity.
-   */
-  @Test
-  public void additionWithFloatQuantityDoesNotWiden() {
-    assertEquals(FloatQuantity.class, ONE_OHM.add(ONE_OHM).getClass());
-  }
-
-  /**
-   * Verifies that addition with DoubleQuantity widens to DoubleQuantity.
-   */
-  @Test
-  public void additionWithDoubleQuantityWidensToDoubleQuantity() {
-    assertEquals(DoubleQuantity.class, ONE_OHM.add(ONE_DOUBLE_OHM).getClass());
-  }
-
-  /**
-   * Verifies that subtraction with FloatQuantity returns a FloatQuantity.
-   */
-  @Test
-  public void subtractionWithFloatQuantityDoesNotWiden() {
-    assertEquals(FloatQuantity.class, ONE_OHM.subtract(ONE_OHM).getClass());
-  }
-
-  /**
-   * Verifies that subtraction with DoubleQuantity widens to DoubleQuantity.
-   */
-  @Test
-  public void subtractionWithDoubleQuantityWidensToDoubleQuantity() {
-    assertEquals(DoubleQuantity.class, ONE_OHM.subtract(ONE_DOUBLE_OHM).getClass());
-  }
-
-  /**
-   * Verifies that multiplication with FloatQuantity returns a FloatQuantity.
-   */
-  @Test
-  public void multiplicationWithFloatQuantityDoesNotWiden() {
-    assertEquals(FloatQuantity.class, ONE_OHM.multiply(ONE_OHM).getClass());
-  }
-
-  /**
-   * Verifies that multiplication with DoubleQuantity widens to DoubleQuantity.
-   */
-  @Test
-  public void multiplicationWithDoubleQuantityWidensToDoubleQuantity() {
-    assertEquals(DoubleQuantity.class, ONE_OHM.multiply(ONE_DOUBLE_OHM).getClass());
-  }
-
-  /**
-   * Verifies that division with FloatQuantity returns a FloatQuantity.
-   */
-  @Test
-  public void divisionWithFloatQuantityDoesNotWiden() {
-    assertEquals(FloatQuantity.class, ONE_OHM.divide(ONE_OHM).getClass());
-  }
-
-  /**
-   * Verifies that division with DoubleQuantity widens to DoubleQuantity.
-   */
-  @Test
-  public void divisionWithDoubleQuantityWidensToDoubleQuantity() {
-    assertEquals(DoubleQuantity.class, ONE_OHM.divide(ONE_DOUBLE_OHM).getClass());
+    assertEquals(-1d, ONE_OHM.negate().getValue());
   }
 }
